Add files via upload
iOS_Oriented_Course_ByteCamp_NEU, assignment 2
This time I tried to write a Student class and completed the task on MacOS.
The Objective-C language is still a bit strange to me, given that I've learned C & Cpp before, the similarities and differences between them can be hard to distinguish
. Still, I sticked to learning the rules in OC language and added some features according to my wishes. The My_Student class provides a function
called "study", which requires a  float parameter, it adds the study_time given in the input and then add it to the member _study_time in My_Student class. Then the info of the student is printed in a certain order. What's more, I realised a construction function for my class, making it possible to construct a class instance almost immediately and conveniently.
I'll learn more in my next class.